Function isolation::isolation::compute

pub fn compute<R, F>(source: &mut BufReader<R>, output: F) where R: Read, F: FnMut(Vec<u64>, usize)

Computes the strongly connected components. Takes a graph stream as an input where the format is as follows:

num_edges
id1 id2 oneway
id1 id2 oneway
...
id1 id2 oneway

where id1 is the start of the edge, id2 is the end of the edge, and oneway is 0 if the edge is bidirectional (id2->id1 is permitted) or 1 if id2 -> id1 is not allowed. num_edges is the number of edges in the file. output is a function that can consume the component vectors and the index of the current component vector.